requirements.txt
aiogram~=2.14
environs~=8.0.0
filters~=1.3.2
mysql-connector~=2.2.9
mysql-connector-python~=8.2.0